Update README.md
Incorporated the installation section into the Readme with a new reformatted look. Specifically, I moved the most common method of installation to the first method, as well as implementing appropriate headers to improve the clarity of the document. For some installation methods, I reformatted each paragraph into steps. However, I may be confused on what are 'steps' and what are 'alternatives,' such as the steps I implemented for the "local installation for pypi." If there are any mistakes you'd like fixed, or if you would like me to edit the Readme, too, please let me know!